Click cancel if a removal request has already been made - closes #5
It seems the UI loads the old or some general dialog before it shows the
final cancel dialog. This is not directly visible but can be seen when
we check the result of the `text()` method.

Adding the timeout in general solves this by waiting for the final /
latest dialog.
A better solution would be to wait until one of the final dialogs
appears and then we trigger a click event on the needed buttons.

The waitFor methods of puppeteer could be some helpful reference in this
case.