Copy and apply the button classes
We can directly copy the button and inner butto classes from the
"request removal" button and save some lines of code.

By doing this we are completely independent of the markup and can safely
copy any styles to the new button and add some required additional
styling like the margin between both buttons.